[backend] refactor: teamDiary 도메인에서 Service 구현체 직접 참조를 인터페이스 기반 참조로 변경하기 위해 Service interface 생성
