+++ backend/src/main/java/com/example/demo/service/TeamDiaryService.java
@@ -1,0 +1,4 @@
+package com.example.demo.service;
+
+public interface TeamDiaryService {
+}
